Исправление и дополнение классов взаимодействующих с БД. Исправление у ID формата на long и всех созависимых классов. Реквизитам таблиц ЦФТ добавлен префик "C_", а таблицам префикс "Z#". Реализован функционал парсера и вставки по наполнению тестовой БД объемными тестовыми данными. Внесены изменения в логику расчета задолженностей. Добавлен метода Plug для тестирования в Main потоке. Функционал расчета задолженностей изменен на паттер Strategy. Увеличино количество логгирования. Добавлен скрипт по установлению конфига.

+++ src/main/java/ru/intel/credits/model/FactOper.java
@@ -17,13 +17,13 @@ public class FactOper {
     private LocalDate date;
     private double summa;
     @EqualsAndHashCode.Include
-    private int oper;
-    private int vidDebt;
-    private int vidDebtDt;
+    private long oper;
+    private long vidDebt;
+    private long vidDebtDt;
     @EqualsAndHashCode.Include
-    private int collectionId;
+    private long collectionId;
 
-    public FactOper(double summa, int oper, int vidDebt, int vidDebtDt, int collectionId) {
+    public FactOper(double summa, long oper, long vidDebt, long vidDebtDt, long collectionId) {
         this.summa = summa;
         this.oper = oper;
         this.vidDebt = vidDebt;
