+++ src/main/java/ru/intel/credits/model/PlanOper.java
@@ -16,14 +16,14 @@ public class PlanOper {
     private double summa;
 
     @EqualsAndHashCode.Include
-    private int oper;
-    private int vidDebt;
-    private int vidDebtDt;
+    private long oper;
+    private long vidDebt;
+    private long vidDebtDt;
 
     @EqualsAndHashCode.Include
-    private int collectionId;
+    private long collectionId;
 
-    public PlanOper(double summa, int oper, int vidDebt, int vidDebtDt, int collectionId) {
+    public PlanOper(double summa, long oper, long vidDebt, long vidDebtDt, long collectionId) {
         this.summa = summa;
         this.oper = oper;
         this.vidDebt = vidDebt;
