+++ src/main/java/ru/intel/credits/repository/CFTRepository.java
@@ -7,19 +7,19 @@ import java.util.concurrent.CopyOnWriteArrayList;
 
 public interface CFTRepository {
 
-    Set<Integer> getIDAllCredInSet();
+    Set<Long> getIDAllCredInSet();
 
-    CopyOnWriteArrayList<Integer> getIDAllCreds();
+    CopyOnWriteArrayList<Long> getIDAllCreds();
 
-    PrCred getCred(int id);
+    PrCred getCred(long id);
 
-    Map<Integer, VidDebt> getAllVidDebts();
+    Map<Long, VidDebt> getAllVidDebts();
 
     Collection<VidOperDog> getAllVidOperDogs();
 
-    Collection<FactOper> getAllFOByCreds(int id);
+    Collection<FactOper> getAllFOByCreds(long id);
 
-    Collection<PlanOper> getAllPOByCreds(int id);
+    Collection<PlanOper> getAllPOByCreds(long id);
 
     Collection<TakeInDebt> getAllTakeInDebt();
 }
